try refactoring source
